pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "optimization-rss"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "feedgen>=0.9",
    "python-dateutil>=2.8",
]

[project.scripts]
update-feed = "optimization_rss.main:main"

[tool.setuptools.packages.find]
where = ["src"]
